[Storage] Move isProofRequired() et al to Marketplace
